backend/requirements.txt
mcp>=1.0.0
python-docx>=1.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
